docs: update based on feedback
