+++ packages/theme/composables/index.ts
@@ -45,3 +45,6 @@ export * from './useMagentoConfiguration';
 
 export * from './types';
 export * from '../modules/GraphQL/types';
+export { default as InjectDOMPurify } from '../plugins/dompurify';
+export { default as InjectFormatCurrency } from '../plugins/fcPlugin';
+export { default as InjectGraphQLClient } from '../plugins/graphqlClient';
