+++ packages/theme/plugins/graphqlClient.ts
@@ -71,3 +71,13 @@ const plugin : Plugin = (ctx, inject) => {
 };
 
 export default plugin;
+
+declare module '@nuxt/types' {
+  interface NuxtAppOptions {
+    $graphql: GraphQlClients;
+  }
+
+  interface Context {
+    $graphql: GraphQlClients;
+  }
+}
